+++ js/constants.js
@@ -1,0 +1,57 @@
+const COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const NAMES = [
+  'Антон',
+  'Максим',
+  'Алёна',
+  'Михаил',
+  'Дмитрий',
+];
+
+const DESCRIPTIONS = [
+  'Около рая',
+  'Указатель',
+  'Вид на море',
+  '(  Y  ) и фотик',
+  'Завтрак',
+  'Black star',
+  'Клубника',
+  'Морс',
+  'Самолёт',
+  'Обувь',
+  'Зной',
+  'Audi',
+  'Перекус',
+  'Котик',
+  'Ноги в тепле',
+  'На высоте',
+  'Хор',
+  'Ретро',
+  'Инновации',
+  'Пальмы',
+  'Салатик',
+  'Закат',
+  'Краб',
+  'Концерт',
+  'Сафари',
+];
+
+const MIN_AVATAR_NUMBER = 1;
+const MAX_AVATAR_NUMBER = 6;
+const MIN_AMOUNT_OF_MESSAGE = 1;
+const MAX_AMOUNT_OF_MESSAGE = 2;
+const MIN_AMOUNT_OF_LIKES = 15;
+const MAX_AMOUNT_OF_LIKES = 200;
+const MIN_ID_OF_COMMENT = 1;
+const MAX_ID_OF_COMMENT = Number.MAX_SAFE_INTEGER;
+const MIN_AMOUNT_OF_COMMENTS = 1;
+const MAX_AMOUNT_OF_COMMENTS = 4;
+
+export {COMMENTS, NAMES, DESCRIPTIONS, MIN_AVATAR_NUMBER, MAX_AVATAR_NUMBER, MIN_AMOUNT_OF_MESSAGE, MAX_AMOUNT_OF_MESSAGE, MIN_AMOUNT_OF_LIKES, MAX_AMOUNT_OF_LIKES, MIN_ID_OF_COMMENT, MAX_ID_OF_COMMENT, MIN_AMOUNT_OF_COMMENTS, MAX_AMOUNT_OF_COMMENTS};
